feature(MaterialsRequest, MaterialsUpdate, MaterialsResponse, Materials) Se modifico el tipo de datos del ancho y alto del material, con el fin de ser mas precisos con los decimales, ya que seran valores en milimetros

+++ src/main/java/com/smartcut/app/Domain/Materials/DTO/MaterialsRequest.java
@@ -26,10 +26,10 @@ public record MaterialsRequest(
         String type,
 
         @NotNull(message = "{message.material.width}")
-        Float width,
+        BigDecimal width,
 
         @NotNull(message = "{message.material.height}")
-        Float height,
+        BigDecimal height,
 
         @NotNull(message = "{message.price.material}")
         @DecimalPesos
